+++ apps/block_scout_web/lib/block_scout_web/api_router.ex
@@ -202,6 +202,7 @@ defmodule BlockScoutWeb.ApiRouter do
     scope "/transactions" do
       get("/", V2.TransactionController, :transactions)
       get("/watchlist", V2.TransactionController, :watchlist_transactions)
+      get("/stats", V2.TransactionController, :stats)
 
       if System.get_env("CHAIN_TYPE") == "polygon_zkevm" do
         get("/zkevm-batch/:batch_number", V2.TransactionController, :polygon_zkevm_batch)
@@ -285,6 +286,7 @@ defmodule BlockScoutWeb.ApiRouter do
       scope "/charts" do
         get("/transactions", V2.StatsController, :transactions_chart)
         get("/market", V2.StatsController, :market_chart)
+        get("/secondary-coin-market", V2.StatsController, :secondary_coin_market_chart)
       end
     end